+++ src/main/java/com/inari/firefly/physics/movement/Integrator.java
@@ -4,8 +4,8 @@ import com.inari.firefly.graphics.ETransform;
 
 public interface Integrator {
     
-    void integrate( final EMovement movement, final ETransform transform, final long deltaTimeInSeconds );
+    void integrate( final EMovement movement, final ETransform transform, final float deltaTimeInSeconds );
     
-    void step( final EMovement movement, final ETransform transform, final long deltaTimeInSeconds );
+    void step( final EMovement movement, final ETransform transform, final float deltaTimeInSeconds );
 
 }
